Extract transport-agnostic proxy core into @varlock/proxy-core
The credential proxy's policy evaluation, substitution guards, placeholder
substitution, response scrubbing, auth-token parsing, and the shared request
pipeline (egress gate -> policy -> cleartext guard -> uninjected-placeholder
guard -> substitution guards -> approval gate -> substitute) now live in a
private workspace package with no node transport dependencies, bundled into
varlock via tsup noExternal. runtime-proxy.ts becomes the node adapter:
servers, MITM, CA handling, upstream identity verification, and streaming
stay put, built on the shared core.

This is the first slice of hosting the proxy on fetch-shaped gateways
(Cloudflare Workers, Vercel), which will reuse the same core.
