+++ .bumpy/proxy-core-extraction.md
@@ -1,0 +1,5 @@
+---
+varlock: patch
+---
+
+internal refactor: proxy policy/substitution/scrubbing logic extracted into a shared transport-agnostic core (no behavior change)
