+++ packages/proxy-core/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@varlock/tsconfig/base.tsconfig.json",
+  "include": ["**/*.ts"],
+  "exclude": ["node_modules", "dist"]
+}
